requirements_docker.txt
azure-core==1.18.0
azure-functions==1.7.2
azure-storage-blob==12.9.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
DateTime==4.3
Deprecated==1.2.13
idna==3.2
isodate==0.6.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
path==16.2.0
path.py==12.5.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
zope.interface==5.4.0
